+++ ci/update-all-dependencies.sh
@@ -7,7 +7,6 @@ set -e
 (cd components/shared_server_code; PATH="$(pwd)/venv/bin:$PATH"; ci/pip-compile.sh; ci/pip-install.sh) &
 (cd components/shared_server_code; PATH="$(pwd)/venv/bin:$PATH"; ci/pip-compile.sh; ci/pip-install.sh) &
 (cd components/external_server; PATH="$(pwd)/venv/bin:$PATH"; ci/pip-compile.sh; ci/pip-install.sh) &
-(cd components/internal_server; PATH="$(pwd)/venv/bin:$PATH"; ci/pip-compile.sh; ci/pip-install.sh) &
 (cd components/collector; PATH="$(pwd)/venv/bin:$PATH"; ci/pip-compile.sh; ci/pip-install.sh) &
 (cd components/notifier; PATH="$(pwd)/venv/bin:$PATH"; ci/pip-compile.sh; ci/pip-install.sh) &
 (cd components/frontend; rm package-lock.json; npm install) &