+++ components/internal_server/src/routes/__init__.py
@@ -1,6 +1,0 @@
-"""Internal routes."""
-
-from .health import get_health
-from .measurement import get_measurements, post_measurement
-from .metric import get_metrics
-from .report import get_reports