+++ tests/feature_tests/src/steps/measurement.py
@@ -3,24 +3,22 @@
 import time
 from datetime import UTC, datetime, timedelta
 
-from asserts import assert_equal, assert_in, assert_true
+from asserts import assert_equal, assert_in
 from behave import given, then, when
 from behave.runner import Context
 from sseclient import SSEClient
 
 
-@when("the collector gets the metrics to measure")
-def get_metrics(context: Context) -> None:
-    """Get the metrics to measure from the server."""
-    context.get("datamodel", internal=True)
-    context.get("metrics", internal=True)
-
-
-@given('the collector has measured "{number}"')
-@when('the collector measures "{number}"')
-@when('the collector measures "{number}" with total "{total}"')
-def measure(context: Context, number: str, total: str = "100") -> None:
-    """Post the measurement."""
+def create_measurement(  # noqa: PLR0913
+    context: Context,
+    status: str | None = None,
+    value: str | None = None,
+    total: str | None = None,
+    issue_id: str | None = None,
+    issue_status_name: str | None = None,
+    parse_error: str | None = None,
+) -> dict:
+    """Create a measurement."""
     entities = []
     if context.table:
         for row in context.table:
@@ -28,77 +26,57 @@ def measure(context: Context, number: str, total: str = "100") -> None:
             entities.append(entity)
     measurement_datetime = datetime.now(tz=UTC)
     measurement_timestamp = measurement_datetime.replace(microsecond=0).isoformat()
-    context.post(
-        "measurements",
-        json={
-            "metric_uuid": context.uuid["metric"],
-            "report_uuid": context.uuid["report"],
-            "has_error": False,
-            "sources": [
-                {
-                    "source_uuid": context.uuid["source"],
-                    "parse_error": None,
-                    "connection_error": None,
-                    "value": number,
-                    "total": total,
-                    "entities": entities,
-                },
-            ],
-            "start": measurement_timestamp,
-            "end": measurement_timestamp,
-        },
-        internal=True,
-    )
+    if issue_id:
+        status_category = "done" if issue_status_name == "Completed" else "todo"
+        issues = [{"issue_id": issue_id, "name": issue_status_name, "status_category": status_category}]
+    else:
+        issues = []
+    return {
+        "metric_uuid": context.uuid["metric"],
+        "report_uuid": context.uuid["report"],
+        "has_error": False,
+        "sources": [
+            {
+                "source_uuid": context.uuid["source"],
+                "parse_error": parse_error,
+                "connection_error": None,
+                "value": value,
+                "total": total,
+                "entities": entities,
+            },
+        ],
+        "count": {"status": status, "value": value},
+        "percentage": {"status": status, "value": value},
+        "version_number": {"status": status, "value": value},
+        "issue_status": issues,
+        "start": measurement_timestamp,
+        "end": measurement_timestamp,
+    }
+
+
+@given('the collector has measured "{number}"')
+@when('the collector measures "{number}"')
+@when('the collector measures "{number}" with status "{status}"')
+@when('the collector measures "{number}" with total "{total}"')
+def measure(context: Context, number: str, status: str = "target_met", total: str = "100") -> None:
+    """Enter a measurement with the given value into the measurement collection in the database."""
+    measurement = create_measurement(context, status, number, total)
+    context.database.measurements.insert_one(measurement)
 
 
 @when("the collector measures issue '{issue_id}' status '{name}'")
-def measure_issue_status(context: Context, issue_id: str, name: str) -> None:
-    """Post the measurement with the issue status."""
-    status_category = "done" if name == "Completed" else "todo"
-    context.post(
-        "measurements",
-        json={
-            "metric_uuid": context.uuid["metric"],
-            "report_uuid": context.uuid["report"],
-            "has_error": False,
-            "sources": [
-                {
-                    "source_uuid": context.uuid["source"],
-                    "parse_error": None,
-                    "connection_error": None,
-                    "value": "10",
-                    "total": "100",
-                    "entities": [],
-                },
-            ],
-            "issue_status": [{"issue_id": issue_id, "name": name, "status_category": status_category}],
-        },
-        internal=True,
-    )
+@when("the collector measures '{number}' with issue '{issue_id}' status '{name}'")
+def measure_issue_status(context: Context, issue_id: str, name: str, number: str = "10") -> None:
+    """Enter a measurement with the issue status into the measurement collection in the database."""
+    measurement = create_measurement(context, "target_met", number, "100", issue_id, name)
+    context.database.measurements.insert_one(measurement)
 
 
 @when("the collector encounters a parse error")
 def parse_error(context: Context) -> None:
-    """Post the parse error."""
-    context.post(
-        "measurements",
-        json={
-            "metric_uuid": context.uuid["metric"],
-            "report_uuid": context.uuid["report"],
-            "has_error": True,
-            "sources": [
-                {
-                    "source_uuid": context.uuid["source"],
-                    "parse_error": "Parse error",
-                    "connection_error": None,
-                    "value": None,
-                    "total": None,
-                    "entities": [],
-                },
-            ],
-        },
-        internal=True,
-    )
+    """Enter a measurement with a parse error into the measurement collection in the database."""
+    measurement = create_measurement(context, parse_error="Parse error")
+    context.database.measurements.insert_one(measurement)
 
 
 @when('the client sets the {attribute} of entity {key} to "{value}"')
@@ -126,12 +104,6 @@ def connect_to_nr_of_measurements_stream(context: Context, stream: str) -> None:
 def skip_update(_context: Context) -> None:
     """Sleep > 10 seconds to give server a chance to skip the next update."""
     time.sleep(10.1)
-
-
-@then("the metric needs to be measured")
-def check_metrics(context: Context) -> None:
-    """Check that the metric needs to be measured."""
-    assert_true(context.uuid["metric"] in context.response.json())
 
 
 @then("the metric {has_or_had} one measurement")